- Tweaked "browse by date, most recent first" behaviour in browse UI.  When the user chooses something like "February 2000" in this case, the browse starts from the *last* date in Feb 2000 (i.e. 29 Feb 2000), instead of the first. If the index looks like this:
March 2000
14 Feb 2000
15 Jan 2000

Entering "Feb 2000" would start the browse at 15 Jan 2000, counter-intuitively,
because "Feb 2000" would appear after "14 Feb 2000" in the index, and the
browse code uses the first available value after that.

Now, in this case, the browse UI will pretend the user wants to start the
browse on "29 Feb 2000", since this means 14 Feb 2000 is used as the starting
point for the browse.

Behaviour when ordering with the oldest item first is unchanged.


git-svn-id: http://scm.dspace.org/svn/repo/trunk@219 9c30dcfa-912a-0410-8fc2-9e0234be79fd