Add the new SWORD module to the 1.5.x branch.  This includes the SWORD code, re-structured to meet the standards already laid out for the 1.5 release (top level project, with sub api and webapp projects).  In addition to this the following additions have been made to the main part of DSpace, which will need to be removed again if SWORD is modularised out again:
- sword-swap-ingest.xsl has been included into config/crosswalks
- sword configuration has been added to dspace.cfg
- loading of the required fields for swap ingest has been added to the main build.xml
- references to dspace-sword have been added to all the relevant pom.xml files in the main for DSpace

Furthermore, this commit includes the source code for the common sword libraries.  These need really to be in the main maven repository, which I will work on as soon as possible before the 1.5 beta release.

This module has been tested to work previously, so it has not been extensively tested again.  Dependency issues in the build system may remain, which have notyet been discovered.

git-svn-id: http://scm.dspace.org/svn/repo/branches/dspace-1_5_x@2598 9c30dcfa-912a-0410-8fc2-9e0234be79fd