Bug Fix for SF# 2013921 - Movement in Submission Workflow Causes Skipped Steps
THIS PATCH WAS ONLY TESTED ON 1.5.x BRANCH.  However, I've updated it for Trunk to hopefully support the new "resource/*" in the URL of the XMLUI submission process.  It's untested, as I couldn't get Trunk's XMLUI to work.

This changes the Submission workflow to use URLs of the form resource/$/submit/*.continue (where the final part is the continuation id).  In addition, the submission process also makes use of Cocoon's Page Local functionality to save the state of individual pages in the submission process.  This ensures that if the user clicks the browser "back button", previous state information can be reloaded (so the flow no longer gets confused as to the current state).

git-svn-id: http://scm.dspace.org/svn/repo/trunk@3021 9c30dcfa-912a-0410-8fc2-9e0234be79fd