Add a certainty factor to line graphs. Uncertain for dateStart, and dateEnd
Added a date to graphs at dateStart and dateEnd (the requested date points),
so that all line graphs would have a data point there on graph, and set to
zero. (Otherwise it would start graph at first entry of data point, which
could be half way later than start of date range). We also set the fake dates
the start/end to be uncertain (dashed lines), and the real data points to
have solid lines.

Also had to refactor chartDataHelper to accept an object, instead of making
signature for that method bigger, uglier and clunkier.